project/requirements_backend.txt

# IDDAW Backend API Requirements
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=9.5.0

# Web framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Authentication
PyJWT>=2.8.0

# Database
# SQLite is included with Python

# Image processing
scikit-image>=0.21.0

# Development
python-dotenv>=1.0.0
waitress>=2.1.2

